backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
Flask-CORS==4.0.0
Flask-Migrate==4.0.4
PyMySQL==1.1.0
python-dotenv==0.21.1
Werkzeug==2.2.3
PyJWT==2.7.0
gunicorn==21.2.0
redis==4.6.0
Flask-Limiter==3.5.0
